Añadida la vista Login si pulsa en index sin estar registrado

+++ src/main/java/com/CRUD/repository/UsuarioRepository.java
@@ -7,7 +7,7 @@ import com.CRUD.entidad.Usuarios;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuarios, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuarios, Integer> {
     
     Optional<Usuarios> findByUsername(String username);//Devuelve Optional<Usuario> para poder manejar que el usuario no exista sin errores.
 }
